+++ lib/danso_function.dart
@@ -1,3 +1,5 @@
 library danso_function;
 
 export './pitch_model/PitchModel.dart';
+export './common/pitch_const/PitchConst.dart';
+export './common/output_pitch_model/OutputPitchModel.dart';