+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release
+* Sorry for users, this might be a package for private application. So you might not as well use this package. Thanks!
+
+
+## 0.0.2
+* PitchModelInterface for danso yulmyeong
+* Made as Singleton.
+* It maps user input frequencies to danso yulmyeong